Module bridge_hub_rococo_runtime::bridge_hub_rococo_config

Bridge definitions that are used on Rococo to bridge with Wococo.

Structs

ActiveOutboundLanesToBridgeHubWococo

BridgeHubRococo

BridgeHubRococo chain from message lane point of view.

BridgeHubRococoUniversalLocation

BridgeHubWococo

BridgeHubWococo chain from message lane point of view.

BridgeHubWococoChainId

BridgeHubWococoMessagesLane

BridgeWococoMessagesPalletInstance

FromAssetHubRococoToAssetHubWococoRoute

MaxUnconfirmedMessagesAtInboundLane

MaxUnrewardedRelayerEntriesAtInboundLane

PriorityBoostPerMessage

StrBridgeRefundBridgeHubWococoMessages

ToBridgeHubWococoXcmBlobHauler

WithBridgeHubWococoMessageBridge

Messaging Bridge configuration for BridgeHubRococo -> BridgeHubWococo

WococoGlobalConsensusNetwork

Constants

DEFAULT_XCM_LANE_TO_BRIDGE_HUB_WOCOCO

Type Aliases

BridgeRefundBridgeHubWococoMessages

Signed extension that refunds relayers that are delivering messages from the Wococo parachain.

FromWococoBridgeHubMessagesProof

Proof of messages, coming from Wococo.

OnBridgeHubRococoBlobDispatcher

Dispatches received XCM messages from other bridge

ToBridgeHubWococoHaulBlobExporter

Export XCM messages to be relayed to the otherside

ToBridgeHubWococoMaximalOutboundPayloadSize

Maximal outbound payload size of BridgeHubRococo -> BridgeHubWococo messages.

ToBridgeHubWococoMessageVerifier

Message verifier for BridgeHubWococo messages sent from BridgeHubRococo

ToWococoBridgeHubMessagesDeliveryProof

Messages delivery proof for Rococo Bridge Hub -> Wococo Bridge Hub messages.