Type Alias bridge_hub_rococo_runtime::bridge_hub_rococo_config::ToWococoBridgeHubMessagesDeliveryProof

pub type ToWococoBridgeHubMessagesDeliveryProof = FromBridgedChainMessagesDeliveryProof<Hash>;

Messages delivery proof for Rococo Bridge Hub -> Wococo Bridge Hub messages.

Aliased Type

struct ToWococoBridgeHubMessagesDeliveryProof {
    pub bridged_header_hash: H256,
    pub storage_proof: Vec<Vec<u8>>,
    pub lane: LaneId,
}

Fields
bridged_header_hash: H256

Hash of the bridge header the proof is for.

storage_proof: Vec<Vec<u8>>

Storage trie proof generated for [Self::bridged_header_hash].

lane: LaneId

Lane id of which messages were delivered and the proof is for.