bridge_hub_rococo_runtime
Module bridge_hub_wococo_config

Bridge definitions that are used on Wococo to bridge with Rococo.

Structs

ActiveOutboundLanesToBridgeHubRococo

BridgeHubRococo

BridgeHubRococo chain from message lane point of view.

BridgeHubRococoChainId

BridgeHubRococoMessagesLane

BridgeHubWococo

BridgeHubWococo chain from message lane point of view.

BridgeHubWococoUniversalLocation

BridgeRococoMessagesPalletInstance

FromAssetHubWococoToAssetHubRococoRoute

MaxUnconfirmedMessagesAtInboundLane

MaxUnrewardedRelayerEntriesAtInboundLane

PriorityBoostPerMessage

RococoGlobalConsensusNetwork

StrBridgeRefundBridgeHubRococoMessages

ToBridgeHubRococoXcmBlobHauler

WithBridgeHubRococoMessageBridge

Messaging Bridge configuration for BridgeHubWococo -> BridgeHubRococo

Constants

DEFAULT_XCM_LANE_TO_BRIDGE_HUB_ROCOCO

Type Aliases

BridgeRefundBridgeHubRococoMessages

Signed extension that refunds relayers that are delivering messages from the Rococo parachain.

FromRococoBridgeHubMessagesProof

Proof of messages, coming from Rococo.

OnBridgeHubWococoBlobDispatcher

Dispatches received XCM messages from other bridge

ToBridgeHubRococoHaulBlobExporter

Export XCM messages to be relayed to the otherside

ToBridgeHubRococoMaximalOutboundPayloadSize

Maximal outbound payload size of BridgeHubWococo -> BridgeHubRococo messages.

ToBridgeHubRococoMessageVerifier

Message verifier for BridgeHubRococo messages sent from BridgeHubWococo

ToRococoBridgeHubMessagesDeliveryProof

Messages delivery proof for Rococo Bridge Hub -> Wococo Bridge Hub messages.