Module bridge_hub_rococo_runtime::bridge_to_bulletin_config

Bridge definitions used on BridgeHubRococo for bridging to Rococo Bulletin.

Rococo Bulletin chain will be the 1:1 copy of the Polkadot Bulletin, so we are reusing Polkadot Bulletin chain primitives everywhere here.

Structs

ActiveLanes

All active routes and their destinations.

ActiveOutboundLanesToRococoBulletin

All active lanes that the current bridge supports.

BridgeRococoToRococoBulletinMessagesPalletInstance

Interior location (relative to this runtime) of the with-RococoBulletin messages pallet.

FromRococoPeopleToRococoBulletinRoute

A route (XCM location and bridge lane) that the Rococo People Chain -> Rococo Bulletin Chain message is following.

MaxUnconfirmedMessagesAtInboundLane

Maximal number of unconfirmed messages at the Rococo Bridge Hub. It matches the maximal number of unconfirmed messages that the single confirmation transaction at Rococo Bulletin Chain may process.

MaxUnrewardedRelayerEntriesAtInboundLane

Maximal number of entries in the unrewarded relayers vector at the Rococo Bridge Hub. It matches the maximal number of unrewarded relayers that the single confirmation transaction at Rococo Bulletin Chain may process.

NeverSentMessage

XCM message that is never sent.

PriorityBoostPerMessage

Priority boost that the registered relayer receives for every additional message in the message delivery transaction.

RococoBulletin

RococoBulletin chain from message lane point of view.

RococoBulletinChainId

Bridge specific chain (network) identifier of the Rococo Bulletin Chain.

RococoBulletinGlobalConsensusNetwork

Rococo Bulletin Network identifier.

RococoBulletinGlobalConsensusNetworkLocation

Relative location of the Rococo Bulletin chain.

RococoPeopleParaId

Identifier of the sibling Rococo People parachain.

RococoPeopleToRococoBulletinMessagesLane

Lane identifier, used to connect Rococo People and Rococo Bulletin chain.

StrOnBridgeHubRococoRefundRococoBulletinMessages

ToRococoBulletinXcmBlobHauler

WithRococoBulletinMessageBridge

Messaging Bridge configuration for BridgeHubRococo -> Rococo Bulletin.

Constants

XCM_LANE_FOR_ROCOCO_PEOPLE_TO_ROCOCO_BULLETIN

Type Aliases

FromRococoBulletinMessagesProof

Proof of messages, coming from Rococo Bulletin chain.

OnBridgeHubRococoRefundRococoBulletinMessages

Signed extension that refunds relayers that are delivering messages from the Rococo Bulletin chain.

ToRococoBulletinHaulBlobExporter

Export XCM messages to be relayed to the other side

ToRococoBulletinMaximalOutboundPayloadSize

Maximal outbound payload size of BridgeHubRococo -> RococoBulletin messages.

ToRococoBulletinMessagesDeliveryProof

Messages delivery proof for Rococo Bridge Hub -> Rococo Bulletin messages.

WithRococoBulletinMessagesInstance

Add XCM messages support for BridgeHubRococo to support Rococo->Rococo Bulletin XCM messages.

XcmOverPolkadotBulletinInstance

Add support for the export and dispatch of XCM programs.