Type Alias bridge_hub_rococo_runtime::bridge_to_bulletin_config::FromRococoBulletinMessagesProof

pub type FromRococoBulletinMessagesProof = FromBridgedChainMessagesProof<Hash>;

Proof of messages, coming from Rococo Bulletin chain.

Aliased Type

struct FromRococoBulletinMessagesProof {
    pub bridged_header_hash: H256,
    pub storage_proof: Vec<Vec<u8, Global>, Global>,
    pub lane: LaneId,
    pub nonces_start: u64,
    pub nonces_end: u64,
}

Fields
bridged_header_hash: H256

Hash of the finalized bridged header the proof is for.

storage_proof: Vec<Vec<u8, Global>, Global>

A storage trie proof of messages being delivered.

lane: LaneId

Messages in this proof are sent over this lane.

nonces_start: u64

Nonce of the first message being delivered.

nonces_end: u64

Nonce of the last message being delivered.