Module bridge_hub_rococo_runtime::bridge_to_westend_config

Bridge definitions used on BridgeHub with the Rococo flavor for bridging to BridgeHubWestend.

Structs

ActiveOutboundLanesToBridgeHubWestend

AssetHubRococoParaId

AssetHubRococoToAssetHubWestendMessagesLane

AssetHubWestendParaId

BridgeHubRococo

BridgeHubRococo chain from message lane point of view.

BridgeHubRococoUniversalLocation

BridgeHubWestend

BridgeHubWestend chain from message lane point of view.

BridgeHubWestendChainId

BridgeRococoToWestendMessagesPalletInstance

CongestedMessage

FromAssetHubRococoToAssetHubWestendRoute

MaxUnconfirmedMessagesAtInboundLane

MaxUnrewardedRelayerEntriesAtInboundLane

PriorityBoostPerMessage

StrOnBridgeHubRococoRefundBridgeHubWestendMessages

ToBridgeHubWestendXcmBlobHauler

UncongestedMessage

WestendGlobalConsensusNetwork

WithBridgeHubWestendMessageBridge

Messaging Bridge configuration for BridgeHubRococo -> BridgeHubWestend

Constants

XCM_LANE_FOR_ASSET_HUB_ROCOCO_TO_ASSET_HUB_WESTEND

Type Definitions

FromWestendBridgeHubMessagesProof

Proof of messages, coming from Westend.

OnBridgeHubRococoRefundBridgeHubWestendMessages

Signed extension that refunds relayers that are delivering messages from the Westend parachain.

ToBridgeHubWestendHaulBlobExporter

Export XCM messages to be relayed to the other side

ToBridgeHubWestendMaximalOutboundPayloadSize

Maximal outbound payload size of BridgeHubRococo -> BridgeHubWestend messages.

ToBridgeHubWestendMessageVerifier

Message verifier for BridgeHubWestend messages sent from BridgeHubRococo

ToWestendBridgeHubMessagesDeliveryProof

Messages delivery proof for Rococo Bridge Hub -> Westend Bridge Hub messages.

WithBridgeHubWestendMessagesInstance

Add XCM messages support for BridgeHubRococo to support Rococo->Westend XCM messages