bridge_hub_rococo_runtime::constants
Module fee

Structs

Perbill

A fixed point representation of a number in the range [0, 1].

ProofSizeToFee

Maps the proof size component of Weight to a fee.

RefTimeToFee

Maps the reference time component of Weight to a fee.

WeightToFee

Handles converting a weight scalar to a fee value, based on the scale and granularity of the node’s balance type.

Constants

TARGET_BLOCK_FULLNESS

The block saturation level. Fees will be updates based on this value.