Type Definition bridge_hub_rococo_runtime::xcm_config::XcmOriginToTransactDispatchOrigin

pub type XcmOriginToTransactDispatchOrigin = (SovereignSignedViaLocation<LocationToAccountId, RuntimeOrigin>, RelayChainAsNative<RelayChainOrigin, RuntimeOrigin>, SiblingParachainAsNative<Origin, RuntimeOrigin>, ParentAsSuperuser<RuntimeOrigin>, SignedAccountId32AsNative<RelayNetwork, RuntimeOrigin>, XcmPassthrough<RuntimeOrigin>);

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.