Crate bridge_hub_rococo_runtime

Modules

api

bridge_hub_rococo_config

Bridge definitions that are used on Rococo to bridge with Wococo.

bridge_hub_wococo_config

Bridge definitions that are used on Wococo to bridge with Rococo.

constants

xcm_config

Structs

BridgeRejectObsoleteHeadersAndMessages

DeliveryRewardInBalance

DepositBase

DepositFactor

ExistentialDeposit

MaxRococoParaHeadDataSize

MaxWococoParaHeadDataSize

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParachainHeadsToKeep

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PotId

RelayChainHeadersToKeep

RelayerStakeLease

RelayerStakeReserveId

RequiredStakeForStakeAndSlash

ReservedDmpWeight

ReservedXcmpWeight

RococoBridgeParachainPalletName

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

TransactionByteFee

Relay Chain TransactionByteFee / 10

Version

WococoBridgeParachainPalletName

Enums

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

Constants

OFFSET

PERIOD

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BridgeGrandpaRococoInstance

Add GRANDPA bridge pallet to track Rococo relay chain on Wococo BridgeHub

BridgeGrandpaWococoInstance

Add GRANDPA bridge pallet to track Wococo relay chain on Rococo BridgeHub

BridgeParachainRococoInstance

Add parachain bridge pallet to track Rococo bridge hub parachain

BridgeParachainWococoInstance

Add parachain bridge pallet to track Wococo bridge hub parachain

BridgeRelayers

BridgeRococoGrandpa

BridgeRococoGrandpaConfig

BridgeRococoMessages

BridgeRococoMessagesConfig

BridgeRococoParachain

BridgeWococoGrandpa

BridgeWococoGrandpaConfig

BridgeWococoMessages

BridgeWococoMessagesConfig

BridgeWococoParachain

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

CumulusXcm

DmpQueue

Executive

Executive: handles dispatch to the various modules.

GenesisConfigDeprecated

Migrations

Migrations to apply on runtime upgrade.

Multisig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

System

SystemConfig

Timestamp

TransactionPayment

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

WithBridgeHubRococoMessagesInstance

Add XCM messages support for BridgeHubWococo to support Wococo->Rococo XCM messages

WithBridgeHubWococoMessagesInstance

Add XCM messages support for BridgeHubRococo to support Rococo->Wococo XCM messages

XcmpQueue