Crate bridge_hub_rococo_runtime

Bridge Hub Rococo Runtime

This runtime currently supports bridging between:

Rococo <> Westend

Rococo <> Rococo Bulletin

Modules

api

bridge_common_config

Bridge definitions that can be used by multiple BridgeHub flavors. All configurations here should be dedicated to a single chain; in other words, we don’t need two chains for a single pallet configuration.

bridge_to_bulletin_config

Bridge definitions used on BridgeHubRococo for bridging to Rococo Bulletin.

bridge_to_ethereum_config

bridge_to_westend_config

Bridge definitions used on BridgeHubRococo for bridging to BridgeHubWestend.

xcm_config

Structs

BaseDeliveryFee

The base fee for the message delivery fees.

BridgeRejectObsoleteHeadersAndMessages

ChainForkVersions

CreateAssetCall

CreateAssetDeposit

DepositBase

DepositFactor

EthereumGatewayAddress

ExistentialDeposit

FeeAssetId

The asset ID for the asset that we use to pay for message delivery fees.

InitStorageVersions

Migration to initialize storage versions for pallets added after genesis.

MaxExecutionHeadersToKeep

MessageQueueServiceWeight

Amount of weight that can be spent per block to service messages. This was increased from 35% to 60% of the max block weight to accommodate the Ethereum beacon light client extrinsics. The force_checkpoint and submit extrinsics (for submit, optionally) includes the sync committee’s pubkeys (512 x 48 bytes)

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Parameters

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PotId

RelayOrigin

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

TransactionByteFee

Relay Chain TransactionByteFee / 10

Version

Enums

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

OFFSET

PERIOD

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BridgePolkadotBulletinGrandpa

BridgePolkadotBulletinGrandpaConfig

BridgePolkadotBulletinMessages

BridgePolkadotBulletinMessagesConfig

BridgeRelayers

BridgeRococoBulletinGrandpa

Proper alias for bridge GRANDPA pallet used to bridge with the bulletin chain.

BridgeRococoBulletinMessages

Proper alias for bridge messages pallet used to bridge with the bulletin chain.

BridgeWestendGrandpa

BridgeWestendGrandpaConfig

BridgeWestendMessages

BridgeWestendMessagesConfig

BridgeWestendParachains

BridgeWestendParachainsConfig

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

CumulusXcm

EthereumBeaconClient

EthereumInboundQueue

EthereumOutboundQueue

EthereumSystem

EthereumSystemConfig

Executive

Executive: handles dispatch to the various modules.

GenesisConfigDeprecated

MessageQueue

Migrations

Migrations to apply on runtime upgrade.

Multisig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

PriceForSiblingParachainDelivery

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

XcmOverBridgeHubWestend

XcmOverPolkadotBulletin

XcmOverRococoBulletin

Proper alias for bridge messages pallet used to bridge with the bulletin chain.

XcmpQueue