Module bridge_hub_rococo_runtime::bridge_hub_wococo_config

Bridge definitions used on BridgeHub with the Wococo flavor.

Structs

ActiveOutboundLanesToBridgeHubRococo

AssetHubWococoParaId

BridgeHubRococo

BridgeHubRococo chain from message lane point of view.

BridgeHubRococoChainId

BridgeHubRococoMessagesLane

BridgeHubWococo

BridgeHubWococo chain from message lane point of view.

BridgeHubWococoUniversalLocation

BridgeWococoToRococoMessagesPalletInstance

CongestedMessage

FromAssetHubWococoToAssetHubRococoRoute

MaxUnconfirmedMessagesAtInboundLane

MaxUnrewardedRelayerEntriesAtInboundLane

PriorityBoostPerMessage

RococoGlobalConsensusNetwork

StrBridgeRefundBridgeHubRococoMessages

ToBridgeHubRococoXcmBlobHauler

UncongestedMessage

WithBridgeHubRococoMessageBridge

Messaging Bridge configuration for BridgeHubWococo -> BridgeHubRococo

Constants

DEFAULT_XCM_LANE_TO_BRIDGE_HUB_ROCOCO

Type Definitions

BridgeRefundBridgeHubRococoMessages

Signed extension that refunds relayers that are delivering messages from the Rococo parachain.

FromRococoBridgeHubMessagesProof

Proof of messages, coming from Rococo.

OnBridgeHubWococoBlobDispatcher

Dispatches received XCM messages from other bridge

OnMessagesDelivered

On messages delivered callback.

ToBridgeHubRococoHaulBlobExporter

Export XCM messages to be relayed to the other side

ToBridgeHubRococoMaximalOutboundPayloadSize

Maximal outbound payload size of BridgeHubWococo -> BridgeHubRococo messages.

ToBridgeHubRococoMessageVerifier

Message verifier for BridgeHubRococo messages sent from BridgeHubWococo

ToRococoBridgeHubMessagesDeliveryProof

Messages delivery proof for RococoBridge Hub -> Wococo BridgeHub messages.

WithBridgeHubRococoMessagesInstance

Add XCM messages support for BridgeHubWococo to support Wococo->Rococo XCM messages